% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_repository.R
\name{as.git_repository}
\alias{as.git_repository}
\title{Open a Git repository, the git2r way}
\usage{
as.git_repository(x, ...)
}
\arguments{
\item{x}{path that is or is in a Git repository; defaults to working
directory}

\item{...}{additional arguments, such as \code{ceiling} from
\code{\link[git2r]{discover_repository}()}}
}
\value{
An S4 \code{\linkS4class{git_repository}} object
}
\description{
\code{githug} uses the \code{\link{git2r}} package, under the hood, to
perform local Git operations. \code{\link{git2r}} handles Git repos as
objects of class \code{\linkS4class{git_repository}}. This function provides
a more flexible version of \code{\link[git2r]{repository}()}, which converts a
path to a \code{\linkS4class{git_repository}}. You might need this for more
exotic Git operations, i.e. to call \code{\link{git2r}} functions that aren't
exposed via \code{githug}.
}
\examples{
\dontrun{
as.git_repository()
as.git_repository("path")
as.git_repository("repo-path/subdir")
as.git_repository("repo-path/subdir", ceiling = 0)

## here's a rather exotic Git operation that githug is unlikely to expose:
## odb_blobs() lists "all blobs reachable from the commits in the object database"
## pre-process the repo with as_git_repository() to prepare for git2r
git2r::odb_blobs(as.git_repository("repo-path"))
}
}

