% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_pat.R
\name{gh_pat}
\alias{gh_pat}
\title{Retrieve GitHub personal access token.}
\usage{
gh_pat(envvar = c("GITHUB_PAT", "GITHUB_TOKEN"))
}
\arguments{
\item{envvar}{Name of the environment variable in which the GitHub personal
access token is stored. Can be a character vector, with names in order of
preference.}
}
\value{
The GitHub personal access token, invisibly.
}
\description{
Consults environment variables for a GitHub personal access token (PAT). By
default, looks for \code{GITHUB_PAT} and \code{GITHUB_TOKEN}, in that order.
Conventions followed by other packages at the time of writing:
\code{\link{git2r}} and \code{devtools} look for \code{GITHUB_PAT},
\code{\link{gh}} looks for \code{GITHUB_TOKEN}. Obtain a PAT from
\url{https://github.com/settings/tokens}. Unless you have a specific reason
to request otherwise, the default scopes are probably sufficient.
}
\details{
How to store your PAT in an environment variable?

\itemize{

\item Identify your home directory. Not sure? Enter
\code{normalizePath("~/")} in the R console.

\item If you don't already have a file here named \code{.Renviron}, create
one. If you have one already, open it for editing.

\item Add a line like this: \code{GITHUB_PAT=blahblahblahblahblahblah}, where
\code{blahblahblahblahblahblah} is your PAT. Make sure the last line in the
file is empty. Otherwise R will \strong{silently} fail to load the file.
Save. And yes you do want to use a filiename that begins with a dot.

\item Restart R. \code{.Renviron} is processed only during
\code{\link{Startup}}.

\item Check your work with \code{Sys.getenv("GITHUB_PAT")}. Your PAT should
print to screen.

}
}
\examples{
## by default, the PAT is not printed to screen
gh_pat()

## if you really want to see it, surround the call with extra parentheses
(gh_pat())
}

