% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_pat-user.R
\name{gh_pat_user}
\alias{gh_pat_user}
\title{Get authenticated user}
\usage{
gh_pat_user(pat = gh_pat())
}
\arguments{
\item{pat}{A GitHub personal access token (PAT).}
}
\value{
character, a GitHub username
}
\description{
Get the authenticated user associated with a GitHub personal access token.
Wraps the
\href{https://developer.github.com/v3/users/#get-the-authenticated-user}{authenticated
user} endpoint of the \href{https://developer.github.com/v3/}{GitHub API}.
}
\examples{
gh_pat_user()
}
\seealso{
\code{\link{gh_pat}()}
}

