% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_pat-username.R
\name{gh_username}
\alias{gh_username}
\title{Get authenticated username}
\usage{
gh_username(pat = gh_pat())
}
\arguments{
\item{pat}{A GitHub personal access token (PAT).}
}
\value{
character, a GitHub username
}
\description{
Get the authenticated username associated with a GitHub personal access
token. Wraps the
\href{https://developer.github.com/v3/users/#get-the-authenticated-user}{authenticated
user} endpoint of the \href{https://developer.github.com/v3/}{GitHub API}.
}
\examples{
gh_username()
}
\seealso{
\code{\link{gh_pat}()}
}

