% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_branch_delete.R
\name{git_branch_delete}
\alias{git_branch_delete}
\title{Delete a branch}
\usage{
git_branch_delete(name, repo = ".")
}
\arguments{
\item{name}{Name of the branch}

\item{repo}{Path to a Git repo. If unspecified, current working directory is
checked to see if it is or is inside a Git repo.}
}
\description{
Delete an existing local branch. You won't be allowed to delete the branch
you're on. Switch to another branch, then delete.
}
\details{
Convenience wrapper around
\code{\link[git2r:branch_delete]{git2r::branch_delete}()}.
}
\examples{
repo <- git_init(tempfile("githug-branches-"))
owd <- setwd(repo)

## commit so that master branch exists
writeLines("Well, we're not in the middle of nowhere...", "nowhere.txt")
git_commit("nowhere.txt", message = "... but we can see it from here.")

## create a branch off of master
git_branch_create("doomed")
git_branch_list()

## switch to doomed branch
git_switch("doomed")
git_branch()

\dontrun{
## try -- and fail -- to delete doomed branch
git_branch_delete("doomed")
}

## switch back to master
git_switch()

## delete the new branch
git_branch_delete("doomed")
git_branch_list()

setwd(owd)
}

