% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_branch.R
\name{git_branch_list}
\alias{git_branch_list}
\title{List branches}
\usage{
git_branch_list(where = c("all", "local", "remote"), repo = ".")
}
\arguments{
\item{where}{Which branches to list: \code{all} (the default), \code{local}
only, or \code{remote} only.}

\item{repo}{Path to a Git repo. If unspecified, current working directory is
checked to see if it is or is inside a Git repo.}
}
\value{
A data frame (or tbl_df) with one row per branch. Variables are
  branch name, type (local vs remote), and a list-column of
  \code{\linkS4class{git_branch}} objects.
}
\description{
Convenience wrapper around \code{\link[git2r]{branches}} from
\code{\link{git2r}}, which coerces the branch list to a data frame.
}
\examples{
## TO DO: come back when I can clone and truly show local v. remote
git_branch_list()
}

