% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_commit.R
\name{git_commit}
\alias{git_commit}
\title{Make a commit}
\usage{
git_commit(..., all = FALSE, force = FALSE, message = NULL, repo = ".")
}
\arguments{
\item{...}{One or more paths or shell glob patterns that will be matched
against files in the repo's working directory. Paths that match will be
added to the set of changes staged for the next commit.}

\item{all}{Logical, consulted if no paths are given. If \code{TRUE},
pre-authorizes the staging of all new files, file deletions, and file
modifications. Emulates \code{git add -A}, which is equivalent to \code{git
add .; git add -u}.}

\item{force}{Logical, defaults to \code{FALSE}. Value \code{TRUE} is required
if any of the to-be-staged paths are currently ignored.}

\item{message}{The commit message. Required. If not supplied, user will get a
chance to provide the message in an interactive session.}

\item{repo}{Path to a Git repo. If unspecified, current working directory is
checked to see if it is or is inside a Git repo.}
}
\value{
SHA of the commit, with a hint about commit time and message stored
  in the \code{hint} attribute
}
\description{
Write file changes to the repository. Convenience wrapper around
\code{\link[git2r:commit]{git2r::commit}()} and, possibly,
\code{\link{git_stage}()}. To stage and commit all current file additions,
deletions, and modifications, use \code{git_commit(all = TRUE)}, which
emulates \code{git add -A && git commit}.
}
\details{
If there are staged changes, the simple call \code{git_commit(message = "my
message")} makes a commit. In the absence of staged changes or when paths are
given via \code{...}, \code{\link{git_stage}()} is called, with arguments
\code{all} and \code{force} also being passed through.
}
\examples{
repo <- git_init(tempfile("githug-"))
owd <- setwd(repo)
write("Are these girls real smart or real real lucky?", "max.txt")
write("You get what you settle for.", "louise.txt")
git_status()
git_commit(all = TRUE,
           message = "Brains'll only get you so far and luck always runs out.")
write("If done properly armed robbery doesn't have to be a totally unpleasant experience.",
      "jd.txt")
git_status()
git_commit("jd.txt", message = "J.D. is charming")
setwd(owd)
}

