% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_config.R
\name{git_config_get}
\alias{git_config_get}
\title{Get Git config variables named in a character vector}
\usage{
git_config_get(vnames = character(), where = c("de_facto", "local",
  "global"), repo = ".")
}
\description{
Get Git config variables named in a character vector
}
\keyword{internal}

