% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_config.R
\name{git_config_list}
\alias{git_config_list}
\title{List Git config variables named in a character vector}
\usage{
git_config_list(vnames = character(), where = c("de_facto", "local",
  "global"), repo = ".")
}
\description{
List Git config variables named in a character vector
}
\keyword{internal}

