% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_config.R
\name{git_config_set}
\alias{git_config_set}
\title{Set Git config variables as given by a named list}
\usage{
git_config_set(vars, where = c("de_facto", "local", "global"), repo = ".")
}
\description{
Set Git config variables as given by a named list
}
\keyword{internal}

