% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_HEAD.R
\name{git_HEAD}
\alias{git_HEAD}
\title{Get HEAD for a repository}
\usage{
git_HEAD(repo = ".")
}
\arguments{
\item{repo}{Path to a Git repo. If unspecified, current working directory is
checked to see if it is or is inside a Git repo.}
}
\value{
A list with components:
\describe{
  \item{\code{branch_name}}{the name and ...}
  \item{\code{branch_type}}{type (local vs. remote) of the branch whose tip
  HEAD currently points to}
  \item{\code{head_sha}}{the SHA and ...}
  \item{\code{head_commit}}{\code{\linkS4class{git_commit}} object for the
  specific commit HEAD points to}
  \item{\code{repo}}{associated \code{\linkS4class{git_repository}} object}
  \item{\code{git_branch}}{associated \code{\linkS4class{git_branch}} object}
}
or \code{NULL} if \code{HEAD} does not exist.
}
\description{
Convenience wrapper around \code{\link[git2r]{head-methods}} from
\code{\link{git2r}}.
}
\examples{
repo <- git_init(tempfile("githug-"))
git_HEAD(repo = repo)
}

