% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_history.R
\name{git_history}
\alias{git_history}
\title{Get the commit history}
\usage{
git_history(repo = ".", ...)
}
\arguments{
\item{repo}{Path to a Git repo. If unspecified, current working directory is
checked to see if it is or is inside a Git repo.}

\item{...}{Optional parameters passed through to
\code{\link[git2r:commits]{git2r::commits}()}. Can include:
\itemize{
\item \code{n} Max number of commits.
\item \code{topological} Logical, requests topological sort, i.e.
parent before child, defaults to \code{TRUE}. Can be combined with
\code{time}.
\item \code{time} Logical, requests chronological sort, defaults to
\code{TRUE}. Can be combined with \code{topological}.
\item \code{reverse} Logical, reverses the order, defaults to
\code{FALSE}.
}}
}
\value{
A data frame with S3 class \code{git_history}, solely for printing
  purposes. Variables: the \code{SHA}, commit \code{message}, \code{when} the
  commit happened, \code{author}, and \code{email}.
}
\description{
Get an overview of the last \code{n} commits. Convenience wrapper around
\code{\link[git2r:commits]{git2r::commits}()}. The print method shows
truncated versions of selected variables, e.g., the commit message, time, and
SHA, but rest assured the full information is present in the returned object.
}
\examples{
repo <- git_init(tempfile("git-history-"))
owd <- setwd(repo)
line1 <- "Thelma: You're a real live outlaw, aren't ya?"
line2 <- paste("J.D.: Well I may be an outlaw, darlin', but you're the one",
               "stealing my heart.")
write(line1, "tl.txt")
git_commit("tl.txt", message = "first commit")
write(line2, "tl.txt", append = TRUE)
git_commit("tl.txt", message = "second commit")
git_history()
setwd(owd)
}

