% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_init.R
\name{git_init}
\alias{git_init}
\title{Create a new Git repository}
\usage{
git_init(path = ".", force = FALSE)
}
\arguments{
\item{path}{Where to create the new Git repo. Defaults to current working
directory. If the \code{path} doesn't exist, it will be created via
\code{dir.create(path, recursive = TRUE)}.}

\item{force}{Whether to create a Git repo, even if it will be inside another
repo. This is generally a bad idea, so use \code{force = TRUE} at your own
risk and consider using Git submodules instead.}
}
\value{
Path to the associated Git repo.
}
\description{
Create a new Git repository
}
\examples{
repo <- tempfile(pattern = "githug-init-example-")
git_init(repo)

## switch working directory to the repo
owd <- setwd(repo)

## Config user and make a commit
git_config(user.name = "thelma", user.email = "thelma@example.org")
writeLines("I don't ever remember feeling this awake.", "thelma.txt")
git_add("thelma.txt")
## TO DO: revisit once I've wrapped add, status, commit, etc.
git2r::status(as_git_repository())
git2r::commit(as_git_repository(), "thelma is awake")
git2r::commits(as_git_repository())[[1]]

setwd(owd)
}

