% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githug_list-class.R
\name{print.githug_list}
\alias{print.githug_list}
\title{Print a list JSON-style}
\usage{
\method{print}{githug_list}(x, ...)
}
\arguments{
\item{x}{The list to print.}

\item{...}{Ignored.}
}
\value{
The input \code{x} is passed through invisibly.
}
\description{
Print a list JSON-style
}

