% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_log.R
\name{git_log}
\alias{git_log}
\title{Get a commit log}
\usage{
git_log(repo = ".")
}
\arguments{
\item{repo}{Path to a Git repo. If unspecified, current working directory is
checked to see if it is or is inside a Git repo.}
}
\value{
A data frame (or tbl_df) with S3 class \code{git_log} for printing
  purposes
}
\description{
Get a commit log. Convenience wrapper around the \code{\link{git2r}}
functions \code{\link[git2r]{coerce-git_repository-method}}, which coerces
the commit log of the repository to a data frame, and
\code{\link[git2r]{commits}}, which returns individual
\code{\linkS4class{git_commit}} objects. The print method shows truncated
versions of selected variables, e.g., the commit message, time, and SHA, but
rest assured the full information is present in the returned object.
}
\examples{
require(dplyr, quietly = TRUE)
repo <- git_init(tempfile("githug-"))
owd <- setwd(repo)
line1 <- "Thelma: You're a real live outlaw, aren't ya?"
line2 <- paste("J.D.: Well I may be an outlaw, darlin', but you're the one",
               "stealing my heart.")
writeLines(line1, "tl.txt")
git_COMMIT(line1)
write(line2, "tl.txt", append = TRUE)
git_COMMIT(line2)
git_log()
setwd(owd)
}

